On x86 (i.e. 32bit) systems, the mpi experiment failed tp display any data (even though it was present in the database) with expview. This was due to the getMetricValues function not being able to match the PMPI_ function addresses recorded by the collector to the function addresses from expanding libmpi.so with dpcl/dyninst.  After much digging in dpcl it was discovered that function addresses are cast to uintptr_t (in daemon/include/SourceObjDABC.h).  To fix this we must first cast function pointers to uintptr_t in the libopenss-runtime function OpenSS_GetAddressOfFunction(). i.e. (uint64_t)((uintptr_t)pointer)

+++ libopenss-runtime/OpenSS_GetAddressOfFunction.c
@@ -51,12 +51,12 @@ uint64_t OpenSS_GetAddressOfFunction(const void* pointer)
 #if defined(__linux) && (defined(__i386) || defined(__x86_64))
 
     /* Return the function pointer directly on Linux/IA32 and Linux/X86-64 */
-    return (uint64_t)pointer;
+    return (uint64_t)((uintptr_t)pointer);
     
 #elif defined(__linux) && defined(__ia64)
     
     /* Return the function pointer from the descriptor on Linux/IA64 */
-    return *((uint64_t*)pointer);
+    return *((uint64_t)((uintptr_t*)pointer));
     
 #else
 #error "Platform/OS Combination Unsupported!"
